added spring cloud api gateway

+++ OrderService/src/main/java/com/rishbootdev/orderservice/dto/OrderRequestDto.java
@@ -11,5 +11,5 @@ public class OrderRequestDto {
 
     private Long id;
     private List<OrderRequestItemDto> items;
-    private BigDecimal totalPrice;
+    private Double totalPrice;
 }
